requirements.txt
mkdocs==0.14.0
mock==1.0.1
nose==1.3.7
rednose==0.4.1
requests==2.7.0
Pygments==2.0.2
python-dateutil== 2.4.1
jinja2==2.7.3
six==1.9.0
